refactor(#123): Phase1 Step1.3-1.8 — split server.py into 5 submodules
dashboard/server.py (1721 行) を責務単位の 5 サブモジュールへ内部分割し、
server.py 自身は全公開シンボルを束ねる re-export shim (228 行) へ縮小した。
公開 API / /api/data JSON contract は byte レベルで不変:

- config.py       : env 解決 + モジュール定数 (区画 A)
- aggregate.py    : period filter + 全 aggregate_* + load_events (区画 B)
- api.py          : build_dashboard_data 集計グルー (区画 C)
- render.py       : テンプレート組み立て + render_static_html (区画 D)
- http_runtime.py : SSE + HTTP server ランタイム (区画 E)

shim は spec_from_file_location 直 import 経路 (test_dashboard.py) のため
サブモジュールを依存順に reload し、import 時 env 評価を patch 済み env 下で
再評価させる。これがないと 2 回目以降の load で env override が効かない。

source-pin テスト 3 件 (sentinel comment / _MAIN_JS_FILES tuple の在処を
grep する characterization テスト) を新ファイルパスへ追従更新。
